all exceptions in one directory
overriding NotFoundInDatabaseException for all not found exceptions

handler for NotFoundInDatabaseException with logs

handler for other exceptions added with logs

code refactoring
