+++ src/main/java/pl/petlovers/Pet2Vet/exceptions/precise_not_found_exceptions/MealNotFoundException.java
@@ -1,0 +1,9 @@
+package pl.petlovers.Pet2Vet.exceptions.precise_not_found_exceptions;
+
+import pl.petlovers.Pet2Vet.exceptions.NotFoundInDatabaseException;
+
+public class MealNotFoundException extends NotFoundInDatabaseException {
+    public MealNotFoundException(Long mealId) {
+        super("Could not find meal with id: " + mealId);
+    }
+}
