+++ src/main/java/pl/petlovers/Pet2Vet/exceptions/precise_not_found_exceptions/NoteNotFoundException.java
@@ -1,0 +1,11 @@
+package pl.petlovers.Pet2Vet.exceptions.precise_not_found_exceptions;
+
+import pl.petlovers.Pet2Vet.exceptions.NotFoundInDatabaseException;
+
+// Exception used to indicate when a note is looked up but not found.
+// thrown, this extra tidbit of Spring MVC configuration is used to render an HTTP 404
+public class NoteNotFoundException extends NotFoundInDatabaseException {
+    public NoteNotFoundException(Long id) {
+        super("Couldn't find note with id = " + id);
+    }
+}
