+++ src/main/java/pl/petlovers/Pet2Vet/meal/MealService.java
@@ -3,6 +3,7 @@ package pl.petlovers.Pet2Vet.meal;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import pl.petlovers.Pet2Vet.exceptions.precise_not_found_exceptions.MealNotFoundException;
 import pl.petlovers.Pet2Vet.pet.Pet;
 import pl.petlovers.Pet2Vet.pet.PetRepository;
 
@@ -26,7 +27,7 @@ public class MealService {
         return petRepository.getById(petId).getMeals();
     }
 
-    public Meal get(long petId, long mealId) {
+    public Meal get(long petId, long mealId) {//todo use pet id to find only specific pet meals
         log.info("Fetching meal with id = " + mealId);
         return mealRepository.findById(mealId)
                 .orElseThrow(() -> new MealNotFoundException(mealId));
