--- src/main/java/pl/petlovers/Pet2Vet/pet/PetNotFoundException.java
@@ -1,7 +1,0 @@
-package pl.petlovers.Pet2Vet.pet;
-
-public class PetNotFoundException extends RuntimeException {
-  public PetNotFoundException(long petId) {
-    super("Could not find pet with id: " + petId);
-  }
-}
