--- src/main/java/pl/petlovers/Pet2Vet/specie/PetSpecieNotFoundException.java
@@ -1,7 +1,0 @@
-package pl.petlovers.Pet2Vet.specie;
-
-public class PetSpecieNotFoundException extends RuntimeException {
-  public PetSpecieNotFoundException(Long specieId) {
-    super("Could not find specie with id: " + specieId);
-  }
-}
