--- src/main/java/pl/petlovers/Pet2Vet/vaccine/VaccineNotFoundException.java
@@ -1,7 +1,0 @@
-package pl.petlovers.Pet2Vet.vaccine;
-
-public class VaccineNotFoundException extends RuntimeException {
-  protected VaccineNotFoundException(Long vaccineId) {
-    super("Could not find vaccine with id: " + vaccineId);
-  }
-}
