+++ src/main/java/pl/petlovers/Pet2Vet/vaccine/VaccineService.java
@@ -5,8 +5,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import pl.petlovers.Pet2Vet.appUser.AppUser;
 import pl.petlovers.Pet2Vet.appUser.AppUserRepository;
+import pl.petlovers.Pet2Vet.exceptions.precise_not_found_exceptions.VaccineNotFoundException;
 import pl.petlovers.Pet2Vet.pet.Pet;
-import pl.petlovers.Pet2Vet.pet.PetNotFoundException;
+import pl.petlovers.Pet2Vet.exceptions.precise_not_found_exceptions.PetNotFoundException;
 
 import java.util.List;
 import java.util.Optional;
@@ -14,9 +15,10 @@ import java.util.Optional;
 @Slf4j
 @Service
 public class VaccineService {
-
   private final AppUserRepository appUserRepository;
   private final VaccineRepository vaccineRepository;
+
+  static final String FETCHING_VACCINE = "Fetching vaccine with id = ";
 
   @Autowired
   public VaccineService(AppUserRepository appUserRepository, VaccineRepository vaccineRepository) {
@@ -30,7 +32,7 @@ public class VaccineService {
   }
 
   public Vaccine get(long vaccineId) {
-    log.info("Fetching vaccine with id = " + vaccineId);
+    log.info(FETCHING_VACCINE + vaccineId);
     return vaccineRepository.findById(vaccineId).orElseThrow(() -> new VaccineNotFoundException(vaccineId));
   }
 
@@ -42,7 +44,7 @@ public class VaccineService {
   }
 
   public Vaccine update(long vaccineId, Vaccine vaccineNewData) {
-    log.info("Fetching vaccine with id = " + vaccineId);
+    log.info(FETCHING_VACCINE + vaccineId);
     Vaccine vaccineFromDB = vaccineRepository.getById(vaccineId);
     log.info("Updating of " + vaccineFromDB + " to " + vaccineNewData.toString());
     vaccineFromDB.modify(vaccineNewData);
@@ -68,7 +70,7 @@ public class VaccineService {
 
   public Vaccine getUserPetsVaccine(long userId, long petId, long vaccineId) {
     Optional<Pet> wantedPet = getUserPetById(userId, petId);
-    log.info("Fetching vaccine with id = " + vaccineId);
+    log.info(FETCHING_VACCINE + vaccineId);
     if (wantedPet.isPresent()) {
       return wantedPet.get().getVaccines().stream()
           .filter(vaccine -> vaccine.getId() == vaccineId)
@@ -114,11 +116,11 @@ public class VaccineService {
 
   private void updateVaccineInPetVaccines(long vaccineId, Vaccine newVaccineData, Optional<Pet> wantedPet) {
     Pet tmpPet = wantedPet.get();
-    log.info("Fetching vaccine with id = " + vaccineId);
-    List<Vaccine> TmpPetVaccines= tmpPet.getVaccines();
-    log.info("Updating of " + TmpPetVaccines.toString() + " to " + newVaccineData.toString());
-    TmpPetVaccines.removeIf(vaccine -> vaccine.getId() == vaccineId);
-    TmpPetVaccines.add(newVaccineData);
+    log.info(FETCHING_VACCINE + vaccineId);
+    List<Vaccine> tmpPetVaccines= tmpPet.getVaccines();
+    log.info("Updating of " + tmpPetVaccines.toString() + " to " + newVaccineData.toString());
+    tmpPetVaccines.removeIf(vaccine -> vaccine.getId() == vaccineId);
+    tmpPetVaccines.add(newVaccineData);
     wantedPet.get().modify(tmpPet);
   }
 
@@ -160,7 +162,7 @@ public class VaccineService {
   }
 
   private AppUser getUser(long userId) {
-    log.info("Fetching user with id = " + userId);
+    log.info(FETCHING_VACCINE + userId);
     return appUserRepository.findById(userId).orElseThrow();
   }
 
@@ -172,7 +174,7 @@ public class VaccineService {
 
   private Optional<Pet> getUserPetById(long userId, long petId){
     List<Pet> userPets = getUsersPets(userId);
-    log.info("Fetching pet with id = " + petId);
+    log.info(FETCHING_VACCINE + petId);
     for (Pet pet : userPets) {
       if (pet.getId() == petId) {
         return Optional.of(pet);
